requirements.txt
discord_webhook
python-dotenv
beautifulsoup4